+++ lib/enums/src/sort_key_product_collection.dart
@@ -1,0 +1,10 @@
+enum SortKeyProductCollection {
+  TITLE,
+  PRICE,
+  BEST_SELLING,
+  CREATED,
+  ID,
+  MANUAL,
+  COLLECTION_DEFAULT,
+  RELEVANCE
+}