+++ lib/graphql_operations/queries/get_all_products_from_collection_by_id.dart
@@ -1,0 +1,86 @@
+const String getCollectionByIdQuery = r'''
+query($id : ID!, $cursor : String){
+  node(id: $id) {
+    ... on Collection {
+      id
+      description
+      descriptionHtml
+      handle
+      title
+      updatedAt
+      image {
+        altText
+        id
+        originalSrc
+      }
+      products(first: 1, sortKey: TITLE, after: $cursor) {
+        edges {
+          cursor
+          node {
+            availableForSale
+            collections(first: 250) {
+              edges {
+                node {
+                  description
+                  descriptionHtml
+                  handle
+                  id
+                  title
+                  updatedAt
+                }
+              }
+            }
+            createdAt
+            description
+            descriptionHtml
+            handle
+            id
+            images(first: 250) {
+              edges {
+                node {
+                  altText
+                  id
+                  originalSrc
+                }
+              }
+            }
+            onlineStoreUrl
+            productType
+            publishedAt
+            tags
+            title
+            updatedAt
+            variants(first: 250) {
+              edges {
+                node {
+                  title
+                  image {
+                    altText
+                    id
+                    originalSrc
+                  }
+                  priceV2 {
+                    amount
+                    currencyCode
+                  }
+                  compareAtPriceV2 {
+                    amount
+                    currencyCode
+                  }
+                  weight
+                  weightUnit
+                  sku
+                  requiresShipping
+                  availableForSale
+                }
+              }
+            }
+          }
+        }
+      pageInfo{
+        hasNextPage
+        }}
+    }
+  }
+}
+''';