+++ lib/graphql_operations/queries/get_all_products_on_query.dart
@@ -1,0 +1,59 @@
+const String getAllProductsOnQueryQuery = r'''
+query( $cursor: String, $sortKey : ProductSortKeys, $query: String){
+  products(query: $query, first: 250, after: $cursor, sortKey: $sortKey) {
+    edges {
+      node {
+        id
+        handle
+        availableForSale
+        createdAt
+        description
+        descriptionHtml
+        onlineStoreUrl
+        productType
+        publishedAt
+        tags
+        title
+        updatedAt
+        images(first: 250) {
+          edges {
+            node {
+              altText
+              id
+              originalSrc
+            }
+          }
+        }
+        variants(first: 250) {
+          edges {
+            node {
+              title
+              image {
+                altText
+                id
+                originalSrc
+              }
+              priceV2 {
+                amount
+                currencyCode
+              }
+              compareAtPriceV2 {
+                amount
+                currencyCode
+              }
+              weight
+              weightUnit
+              availableForSale
+              sku
+              requiresShipping
+            }
+          }
+        }
+      }
+      cursor
+    }
+    pageInfo {
+      hasNextPage
+    }
+  }
+}''';