+++ lib/models/src/article.dart
@@ -7,13 +7,13 @@ class Articles {
 
   static Articles fromJson(Map<String, dynamic> json){
     return Articles(
-        articleList: _getArticleList(json)
+        articleList: _getArticleList(json ?? const {})
     );
   }
 
   static _getArticleList(Map<String, dynamic> json) {
     List<Article> articleList = [];
-    json['edges'].forEach((article) => articleList.add(Article.fromJson(article)) );
+    json['edges']?.forEach((article) => articleList.add(Article.fromJson(article ?? const {})) );
     return articleList;
   }
 }
@@ -37,17 +37,17 @@ class Article {
 
   static Article fromJson(Map<String, dynamic> json){
     return Article(
-      author: AuthorV2.fromJson((json['node'] ?? const {})['authorV2']),
-      commentList: _getCommentList((json['node'] ?? const {})['comments']),
+      author: AuthorV2.fromJson(((json['node'] ?? const {})['authorV2']) ?? const {}),
+      commentList: _getCommentList((json['node'] ?? const {})['comments'] ?? const {}),
       content: (json['node'] ?? const {})['content'],
       contentHtml: (json['node'] ?? const {})['contentHtml'],
       excerpt: (json['node'] ?? const {})['excerpt'],
       excerptHtml: (json['node'] ?? const {})['excerptHtml'],
       handle: (json['node'] ?? const {})['handle'],
       id: (json['node'] ?? const {})['id'],
-      image: ShopifyImage.fromJson((json['node'] ?? const {})['image']),
+      image: ShopifyImage.fromJson((json['node'] ?? const {})['image'] ?? const {}),
       publishedAt: (json['node'] ?? const {})['publishedAt'],
-      tags: _getTagsList(json),
+      tags: _getTagsList(json ?? const {}),
       title: (json['node'] ?? const {})['title'],
       url: (json['node'] ?? const {})['url'],
     );
@@ -55,13 +55,13 @@ class Article {
 
   static _getCommentList(Map<String, dynamic> json){
     List<Comment> commentList = [];
-    json['edges'].forEach((comment) => commentList.add(Comment.fromJson(comment)));
+    json['edges']?.forEach((comment) => commentList.add(Comment.fromJson(comment ?? const {})));
     return commentList;
   }
 
   static _getTagsList(Map<String, dynamic> json){
     List<String> tagsList = [];
-    (json['node'] ?? const {})['tags'].forEach((tag) => tagsList.add(tag));
+    (json['node'] ?? const {})['tags']?.forEach((tag) => tagsList.add(tag));
     return tagsList;
   }
 }