+++ lib/models/src/product.dart
@@ -1,22 +1,21 @@
 class Products {
-final List<Product> productList;
-final bool hasNextPage;
+  final List<Product> productList;
+  final bool hasNextPage;
 
   Products({this.productList, this.hasNextPage});
 
-  static Products fromJson(Map<String, dynamic> json){
+  static Products fromJson(Map<String, dynamic> json) {
     return Products(
-      productList: _getProductList(json ?? const {}),
-      hasNextPage: (json['pageInfo'] ?? const {})['hasNextPage']
-    );
+        productList: _getProductList(json ?? const {}),
+        hasNextPage: (json['pageInfo'] ?? const {})['hasNextPage']);
   }
 
-static List<Product> _getProductList(Map<String, dynamic> json) {
-  List<Product> productList = [];
-  json['edges']?.forEach((e) => productList.add(Product.fromJson(e)));
-  return productList;
-}
-
+  static List<Product> _getProductList(Map<String, dynamic> json) {
+    List<Product> productList = [];
+    json['edges']
+        ?.forEach((e) => productList.add(Product.fromJson(e ?? const {})));
+    return productList;
+  }
 }
 
 class Product {
@@ -51,60 +50,57 @@ class Product {
       this.publishedAt,
       this.tags,
       this.updatedAt,
-      this.cursor
-      });
+      this.cursor});
 
   static Product fromJson(Map<String, dynamic> json) {
     return Product(
-      collectionList: _getCollectionList(json),
-      id: (json['node'] ?? const {})['id'],
-      title: (json['node'] ?? const {})['title'],
-      availableForSale:
-          (json['node'] ?? const {})['availableForSale'],
-      createdAt: (json['node'] ?? const {})['createdAt'],
-      description: (json['node'] ?? const {})['description'],
-      productVariants: _getProductVariants(json),
-      descriptionHtml: (json['node'] ?? const {})['descriptionHtml'],
-      handle: (json['node'] ?? const {})['handle'],
-      onlineStoreUrl: (json['node'] ?? const {})['onlineStoreUrl'],
-      productType: (json['node'] ?? const {})['productType'],
-      publishedAt: (json['node'] ?? const {})['publishedAt'],
-      tags: _getTags(json),
-      updatedAt: (json['node'] ?? const {})['updatedAt'],
-      cursor: json['cursor']
-    );
+        collectionList: _getCollectionList(json ?? const {}),
+        id: (json['node'] ?? const {})['id'],
+        title: (json['node'] ?? const {})['title'],
+        availableForSale: (json['node'] ?? const {})['availableForSale'],
+        createdAt: (json['node'] ?? const {})['createdAt'],
+        description: (json['node'] ?? const {})['description'],
+        productVariants: _getProductVariants(json ?? const {}),
+        descriptionHtml: (json['node'] ?? const {})['descriptionHtml'],
+        handle: (json['node'] ?? const {})['handle'],
+        onlineStoreUrl: (json['node'] ?? const {})['onlineStoreUrl'],
+        productType: (json['node'] ?? const {})['productType'],
+        publishedAt: (json['node'] ?? const {})['publishedAt'],
+        tags: _getTags(json ?? const {}),
+        updatedAt: (json['node'] ?? const {})['updatedAt'],
+        cursor: json['cursor']);
   }
 
   static List<ProductVariant> _getProductVariants(Map<String, dynamic> json) {
     List<ProductVariant> productVariants = [];
     (((json['node'] ?? const {})['variants'] ?? const {})['edges'] ?? const [])
-        .forEach((v) {
-      productVariants.add(ProductVariant.fromJson(v.data));
+        ?.forEach((v) {
+      if (v?.data != null)
+        productVariants.add(ProductVariant.fromJson(v?.data ?? const {}));
     });
     return productVariants;
   }
 
   static List<String> _getTags(Map<String, dynamic> json) {
     List<String> tags = [];
-    (json['node'] ?? const {})['tags']?.forEach((e) => tags.add(e));
+    (json['node'] ?? const {})['tags']?.forEach((e) => tags.add(e ?? const {}));
     return tags;
   }
 
-  static List<AssociatedCollections> _getCollectionList(Map<String, dynamic> json) {
+  static List<AssociatedCollections> _getCollectionList(
+      Map<String, dynamic> json) {
     List<AssociatedCollections> collectionList = [];
-    (((json['node'] ?? const {})['collections'] ?? const {})['edges'] ?? const [])
-        .forEach((v) {
-      collectionList.add(AssociatedCollections.fromJson(v.data));
-  });
-  return collectionList;
+    (((json['node'] ?? const {})['collections'] ?? const {})['edges'] ??
+            const [])
+        ?.forEach((v) {
+      if (v?.data != null)
+        collectionList.add(AssociatedCollections.fromJson(v?.data ?? const {}));
+    });
+    return collectionList;
+  }
 }
-
-}
-
-
 
 class AssociatedCollections {
-
   final String description;
   final String descriptionHtml;
   final String id;
@@ -112,17 +108,22 @@ class AssociatedCollections {
   final String updatedAt;
   final String title;
 
-  AssociatedCollections({this.description, this.descriptionHtml, this.id, this.handle, this.updatedAt, this.title});
+  AssociatedCollections(
+      {this.description,
+      this.descriptionHtml,
+      this.id,
+      this.handle,
+      this.updatedAt,
+      this.title});
 
-  static AssociatedCollections fromJson(Map<String, dynamic> json){
+  static AssociatedCollections fromJson(Map<String, dynamic> json) {
     return AssociatedCollections(
-      description: (json['node'] ?? const {})['description'],
-      descriptionHtml: (json['node'] ?? const {})['descriptionHtml'],
-      id: (json['node'] ?? const {})['id'],
-      handle: (json['node'] ?? const {})['handle'],
-      updatedAt: (json['node'] ?? const {})['updatedAt'],
-      title: (json['node'] ?? const {})['title']
-    );
+        description: (json['node'] ?? const {})['description'],
+        descriptionHtml: (json['node'] ?? const {})['descriptionHtml'],
+        id: (json['node'] ?? const {})['id'],
+        handle: (json['node'] ?? const {})['handle'],
+        updatedAt: (json['node'] ?? const {})['updatedAt'],
+        title: (json['node'] ?? const {})['title']);
   }
 }
 
@@ -170,7 +171,8 @@ class ProductVariant {
       price:
           PriceV2.fromJson((json['node'] ?? const {})['priceV2'] ?? const {}),
       title: (json['node'] ?? const {})['title'],
-      image: ShopifyImage.fromJson((json['node'] ?? const {})['image'] ?? {}),
+      image: ShopifyImage.fromJson(
+          (json['node'] ?? const {})['image'] ?? const {}),
       compareAtPrice: PriceV2.fromJson(
           (json['node'] ?? const {})['compareAtPriceV2'] ?? const {}),
       weight: (json['node'] ?? const {})['weight'],